Simplify finalizer check
To make finalizers work for WorkLoaders that might route things asynchronously, we need to store a boolean to say if work has a finalizer.

Before we were checking the IRunnable at time of routing the finalizer, but that might be run from a totally different service that doesn't have access to the work. We don't want to route the finalizer if there isn't one, so this boolean saves us the check.